renames stdlib sub-package to common
