+++ arcana/core/cli/deploy.py
@@ -41,7 +41,7 @@ containing multiple specifications
 TARGET is the type of image to build, e.g. arcana.xnat.deploy:XnatApp
 the target should resolve to a class deriviing from arcana.core.deploy.App.
 If it is located under the `arcana.deploy`, then that prefix can be dropped, e.g.
-stdlib:App
+common:App
 """,
 )
 @click.argument("spec_root", type=click.Path(exists=True, path_type=Path))
