+++ arcana/core/cli/tests/test_cli_deploy.py
@@ -47,7 +47,7 @@ def test_deploy_make_app_cli(command_spec, cli_runner, work_dir):
         make_app,
         [
             str(spec_path),
-            "stdlib:App",
+            "common:App",
             "--build-dir",
             str(build_dir),
             "--registry",
@@ -90,7 +90,7 @@ def test_deploy_remake_app_cli(command_spec, docker_registry, cli_runner, run_pr
             make_app,
             [
                 str(spec_path),
-                "stdlib:App",
+                "common:App",
                 "--build-dir",
                 str(build_dir),
                 "--registry",
@@ -180,7 +180,7 @@ description: >-
   a test of the YAML join functionality
 command:
   task: arcana.testing.tasks:identity_file
-  row_frequency: stdlib:Samples[sample]
+  row_frequency: common:Samples[sample]
   inputs:
     in_file:
       datatype: text/text-file
@@ -298,7 +298,7 @@ command:
         field: flags
         datatype: str
         help_string: description of flags param
-    row_frequency: stdlib:Clinical[session]
+    row_frequency: common:Clinical[session]
     configuration:
       executable: /usr/local/miniconda/bin/mriqc
       dataset: /work/bids-dataset
