+++ arcana/core/deploy/tests/test_licenses.py
@@ -45,7 +45,7 @@ def test_buildtime_license(license_file, run_prefix: str, work_dir: Path, cli_ru
         make_app,
         args=[
             str(root_dir),
-            "stdlib:App",
+            "common:App",
             "--build-dir",
             str(build_dir),
             "--license",
